feat: 포의 궁성 내부 영역에서의 이동 검증 기능 추가
