+++ src/main/java/janggi/board/strategy/NormalPlaceStrategy.java
@@ -1,7 +1,7 @@
 package janggi.board.strategy;
 
 import janggi.board.Position;
-import janggi.piece.Canon;
+import janggi.piece.Cannon;
 import janggi.piece.Chariot;
 import janggi.piece.Elephant;
 import janggi.piece.General;
@@ -29,7 +29,7 @@ public class NormalPlaceStrategy implements PlaceStrategy {
         Team team = Team.RED;
 
         placePieces(board, team, Soldier::new, 6, 0, 2, 4, 6, 8);
-        placePieces(board, team, Canon::new, 7, 1, 7);
+        placePieces(board, team, Cannon::new, 7, 1, 7);
         board.put(new Position(4, 8), new General(team));
         placePieces(board, team, Chariot::new, 9, 0, 8);
         placePieces(board, team, Elephant::new, 9, 1, 6);
@@ -41,7 +41,7 @@ public class NormalPlaceStrategy implements PlaceStrategy {
         Team team = Team.GREEN;
 
         placePieces(board, team, Soldier::new, 3, 0, 2, 4, 6, 8);
-        placePieces(board, team, Canon::new, 2, 1, 7);
+        placePieces(board, team, Cannon::new, 2, 1, 7);
         board.put(new Position(4, 1), new General(team));
         placePieces(board, team, Chariot::new, 0, 0, 8);
         placePieces(board, team, Elephant::new, 0, 1, 6);
